# pulse/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_whg517_node-pulse_pulse_internal_models.CreateNodeRequest:
    properties:
      ip:
        maxLength: 45
        type: string
      name:
        maxLength: 255
        type: string
      region:
        maxLength: 100
        type: string
      tags:
        additionalProperties: true
        type: object
    required:
    - ip
    - name
    - region
    type: object
...
  github_com_whg517_node-pulse_pulse_internal_models.DeleteNodeResponse:
    properties:
      message:
        type: string
      timestamp:
        type: string
    type: object
...
  github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  github_com_whg517_node-pulse_pulse_internal_models.GetNodeStatusResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.NodeStatusData'
      message:
        type: string
      timestamp:
        type: string
    type: object
...
  github_com_whg517_node-pulse_pulse_internal_models.Node:
    properties:
      created_at:
        type: string
      id:
        type: string
      ip:
        type: string
      last_heartbeat:
        description: Beacon heartbeat time
        type: string
      last_report_time:
        description: Data write time
        type: string
      name:
        type: string
      region:
        type: string
      status:
        description: online/offline/connecting
        type: string
      tags:
        description: JSONB stored as string
        type: string
      updated_at:
        type: string
    type: object
  github_com_whg517_node-pulse_pulse_internal_models.NodeStatus:
    properties:
      id:
        type: string
      last_heartbeat:
        type: string
      last_report_time:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
  github_com_whg517_node-pulse_pulse_internal_models.NodeStatusData:
    properties:
      node:
        $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.NodeStatus'
    type: object
...
  github_com_whg517_node-pulse_pulse_internal_models.UpdateNodeRequest:
    properties:
      ip:
        maxLength: 45
        type: string
      name:
        maxLength: 255
        type: string
      region:
        maxLength: 100
        type: string
      tags:
        additionalProperties: true
        type: object
    type: object
  github_com_whg517_node-pulse_pulse_internal_models.UpdateNodeResponse:
    properties:
      data:
        properties:
          node:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.Node'
        type: object
      message:
        type: string
      timestamp:
        type: string
    type: object
...
host: localhost:6532
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/kevin/node-pulse
  description: |-
    Node Pulse is a distributed monitoring system for tracking node health, metrics, and alerts in real-time.

    ## Features
    - Real-time beacon heartbeat monitoring
    - Multi-metric data collection and analysis
    - Configurable alert rules with suppression
    - Webhook notifications for alerts
    - Historical data queries and export
    - Role-based access control (RBAC)

    ## Authentication
    Most endpoints require authentication. Use the `/api/v1/auth/login` endpoint to obtain a session token.
    Include the token in the `Authorization` header: `Bearer <your-token>`
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Node Pulse API
  version: "1.0"
paths:
...
  /nodes:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all nodes. Supports filtering by region.
      parameters:
      - description: Filter by region
        example: us-west
        in: query
        name: region
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of nodes
          schema:
            items:
              $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.Node'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all nodes
      tags:
      - nodes
    post:
      consumes:
      - application/json
      description: Creates a new monitoring node. If a node with the same name and
        IP exists, it will be updated instead.
      parameters:
      - description: Node creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.CreateNodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Node updated successfully (duplicate found)
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.Node'
        "201":
          description: Node created successfully
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.Node'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "403":
          description: Forbidden (requires admin or operator role)
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new node
      tags:
      - nodes
  /nodes/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a node. Requires confirmation query parameter and admin
        or operator role.
      parameters:
      - description: Node UUID
        in: path
        name: id
        required: true
        type: string
      - description: Must be 'true' to confirm deletion
        in: query
        name: confirm
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Node deleted successfully
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.DeleteNodeResponse'
        "400":
          description: Invalid request or confirmation required
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "403":
          description: Forbidden (requires admin or operator role)
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "404":
          description: Node not found
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a node
      tags:
      - nodes
    get:
      consumes:
      - application/json
      description: Retrieves detailed information about a specific node by its UUID.
      parameters:
      - description: Node UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Node details
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.Node'
        "400":
          description: Invalid UUID format
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "404":
          description: Node not found
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get node by ID
      tags:
      - nodes
    put:
      consumes:
      - application/json
      description: Updates an existing node. Requires admin or operator role.
      parameters:
      - description: Node UUID
        in: path
        name: id
        required: true
        type: string
      - description: Node update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.UpdateNodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Node updated successfully
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.UpdateNodeResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "403":
          description: Forbidden (requires admin or operator role)
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "404":
          description: Node not found
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a node
      tags:
      - nodes
  /nodes/{id}/status:
    get:
      consumes:
      - application/json
      description: Retrieves the current operational status of a node including recent
        metrics.
      parameters:
      - description: Node UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Node status
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.GetNodeStatusResponse'
        "400":
          description: Invalid UUID format
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "404":
          description: Node not found
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_whg517_node-pulse_pulse_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get node status
      tags:
      - nodes
...
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"